+++ tests/test_spdx.py
@@ -1,0 +1,33 @@
+# Copyright Kevin Deldycke <kevin@deldycke.com> and contributors.
+#
+# This program is Free Software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+from __future__ import annotations
+
+import pytest
+
+from meta_package_manager.spdx import normalize_spdx_id
+
+
+@pytest.mark.parametrize(
+    ("raw_str", "expected"),
+    (
+        ("SPDXRef-Package-brew-openjdk@11", "SPDXRef-Package-brew-openjdk-11"),
+        ("SPDXRef-my.Super.package", "SPDXRef-my.Super.package"),
+        ("SPDXRef-my---Super.package", "SPDXRef-my-Super.package"),
+    ),
+)
+def test_normalize_spdx_id(raw_str, expected):
+    assert normalize_spdx_id(raw_str) == expected